[refactor] move virtual object from scene
