+++ include/zen/virtual-object.h
