+++ zna/virtual-object.c
