+++ zna/virtual-object/gl-texture.c
